cbrn_safe_eval/requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pandas>=2.2.2
pydantic>=2.8.2
tenacity>=8.5.0
tqdm>=4.66.4
requests>=2.32.3
rich>=13.7.1
pytest>=8.3.3

